requirements.txt
black==20.8b1
isort==5.4.2
mypy==0.982
pylint==2.6.0
pytest==6.0.1
py~=1.11.0
colorama~=0.4.5
future~=0.18.2
Pillow~=9.2.0
pip~=22.2.2
attrs~=22.1.0
toml~=0.10.2
wheel~=0.37.1
Jinja2~=3.1.2
toolz~=0.12.0
numpy~=1.23.4
blis~=0.7.8
tomli~=2.0.1
pytz~=2022.4
tqdm~=4.52.0
requests~=2.24.0
cymem~=2.0.7
setuptools~=60.2.0
Cython~=0.29.28
pathy~=0.6.2
typer~=0.4.2
spacy~=3.4.1
scipy~=1.9.2
matplotlib~=3.3.3
thinc~=8.1.4
srsly~=2.4.4
wasabi~=0.10.1
click~=8.1.3
catalogue~=2.0.8
pydantic~=1.9.2
packaging~=21.3
langcodes~=3.3.0
preshed~=3.0.8
murmurhash~=1.0.9
confection~=0.0.3
gensim~=4.2.0
MarkupSafe~=2.1.1
joblib~=1.2.0
pandas~=1.1.4
python-dateutil~=2.8.2
numexpr~=2.8.3
pygame~=2.1.3.dev6
astroid~=2.5
mccabe~=0.6.1
pluggy~=0.13.1
iniconfig~=1.1.1
atomicwrites~=1.4.1
wrapt~=1.12.1
certifi~=2022.9.24
chardet~=3.0.4
cytoolz~=0.12.0
scikit-learn~=1.1.2
threadpoolctl~=3.1.0
networkx~=2.5
textacy~=0.12.0
cachetools~=5.2.0
jellyfish~=0.9.0
pyphen~=0.13.0
urllib3~=1.25.11
six~=1.16.0
decorator~=5.1.1
funcy~=1.17
pyLDAvis~=2.1.2
idna~=2.10
pyparsing~=3.0.9
GameOfLife~=0.1.3
cycler~=0.11.0
kiwisolver~=1.4.4
responses~=0.12.1
httpretty~=1.0.2
python-louvain~=0.14